+++ CHANGELOG.md
@@ -11,6 +11,7 @@ You can also check the
 
 # Unreleased
 
+
 - Features
   - Added option for hiding legend titles using a toggle switch
 - Fixes
@@ -28,6 +29,8 @@ You can also check the
 - Maintenance
   - Added authentication method to e2e tests
   - Added authentication to vercel previews for easier testing
+  - flags can now be toggle by clicking "ctrl + t + d" on the keyboard (t,d =>
+    toggle debug)
 
 # [5.2.4] - 2025-02-06